extract sources component
