+++ frontend/src/cards/ui/MultiMapDialog.module.scss
@@ -2,3 +2,22 @@
   width: 300px;
   padding: 15px;
 }
+
+.SmallMultipleLegendMap {
+  width: 300px;
+  padding: 15px;
+  text-align: center;
+}
+
+.LegendDiv {
+  width: 120px;
+  margin: auto;
+}
+
+.FooterButtonContainer {
+  float: right;
+}
+
+.FooterSourcesContainer {
+  padding: 7px;
+}
