+++ frontend/src/cards/ui/MultiMapDialog.tsx
@@ -2,15 +2,17 @@ import React from "react";
 import Alert from "@material-ui/lab/Alert";
 import Button from "@material-ui/core/Button";
 import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
 import DialogContent from "@material-ui/core/DialogContent";
 import { Grid } from "@material-ui/core";
 import { ChoroplethMap } from "../../charts/ChoroplethMap";
 import { Fips } from "../../data/utils/Fips";
 import { Legend } from "../../charts/Legend";
+import { MapOfDatasetMetadata } from "../../data/utils/DatasetTypes";
 import { MetricConfig } from "../../data/config/MetricConfig";
 import { Row, FieldRange } from "../../data/utils/DatasetTypes";
+import { Sources } from "./Sources";
 import styles from "./MultiMapDialog.module.scss";
+import { MetricQueryResponse } from "../../data/query/MetricQuery";
 
 const QUANTILE = "quantile";
 const VEGA_LEGEND_REFERENCE_LINK =
@@ -33,6 +35,10 @@ export interface MultiMapDialogProps {
   open: boolean;
   // Closes the dialog in the parent component
   handleClose: () => void;
+  // Dataset IDs required the source footer
+  queryResponses: MetricQueryResponse[];
+  // Metadata required for the source footer
+  metadata: MapOfDatasetMetadata;
 }
 
 /*
@@ -50,24 +56,29 @@ export function MultiMapDialog(props: MultiMapDialogProps) {
     >
       <DialogContent dividers={true}>
         <Grid container justify="space-around">
-          <Grid item xs={6}>
-            <Alert severity="info">
+          <Grid item xs={12}>
+            <Alert severity="warning">
+              <b>Compare with caution</b>
+              <br />
               This scale is a{" "}
               <a href={VEGA_LEGEND_REFERENCE_LINK}>{QUANTILE}</a> scale,
               optimized for visualizing and comparing across demographics.
             </Alert>
           </Grid>
-          <Grid item xs={6}>
-            <Legend
-              metric={props.metricConfig}
-              legendTitle={props.metricConfig.fullCardTitleName}
-              legendData={props.data}
-              scaleType={QUANTILE}
-              sameDotSize={true}
-            />
-          </Grid>
         </Grid>
         <Grid container justify="space-around">
+          <Grid item className={styles.SmallMultipleLegendMap}>
+            <b>Legend</b>
+            <div className={styles.LegendDiv}>
+              <Legend
+                metric={props.metricConfig}
+                legendTitle={props.metricConfig.fullCardTitleName}
+                legendData={props.data}
+                scaleType={QUANTILE}
+                sameDotSize={true}
+              />
+            </div>
+          </Grid>
           {props.breakdownValues.map((breakdownValue) => {
             const dataForValue = props.data.filter(
               (row: Row) => row[props.breakdown] === breakdownValue
@@ -96,11 +107,19 @@ export function MultiMapDialog(props: MultiMapDialogProps) {
           })}
         </Grid>
       </DialogContent>
-      <DialogActions>
-        <Button onClick={props.handleClose} color="primary">
-          Close
-        </Button>
-      </DialogActions>
+      <div>
+        <div className={styles.FooterButtonContainer}>
+          <Button onClick={props.handleClose} color="primary">
+            Close
+          </Button>
+        </div>
+        <div className={styles.FooterSourcesContainer}>
+          <Sources
+            queryResponses={props.queryResponses}
+            metadata={props.metadata}
+          />
+        </div>
+      </div>
     </Dialog>
   );
 }
